Set title text on episode increment button.
I had to fire up the inspector to be sure this button did what I thought
it did; to me it looked like it was more to do with the title of the
show than the episode counter. Had there been title text, I would have
been a lot surer a lot sooner.